+++ package.json
@@ -10,6 +10,7 @@
     "firebase": "^10.1.0",
     "formik": "^2.4.3",
     "papaparse": "^5.4.1",
+    "point-in-polygon": "^1.1.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
